feat: rewrite dump_info
